move popover into component
